backend/requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Celery
celery>=5.3,<6.0
django-celery-beat>=2.5,<3.0
django-celery-results>=2.5,<3.0

# Database
psycopg2-binary>=2.9,<3.0

# HTTP Client
requests>=2.31,<3.0

# Production server
gunicorn>=21.0,<22.0

# Development & Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<5.0

# Code quality
flake8>=6.1,<7.0
black>=23.0,<24.0
isort>=5.12,<6.0

# Type hints
mypy>=1.5,<2.0
django-stubs>=4.2,<5.0
